+++ pymongo/database.py
@@ -142,7 +142,7 @@ class Database(common.BaseObject):
 
     @property
     def system_js(self):
-        """A :class:`SystemJS` helper for this :class:`Database`.
+        """**DEPRECATED**: :class:`SystemJS` helper for this :class:`Database`.
 
         See the documentation for :class:`SystemJS` for more details.
         """
@@ -990,7 +990,7 @@ class Database(common.BaseObject):
 
     def authenticate(self, name=None, password=None,
                      source=None, mechanism='DEFAULT', **kwargs):
-        """Authenticate to use this database.
+        """**DEPRECATED**: Authenticate to use this database.
 
         Authentication lasts for the life of the underlying client
         instance, or until :meth:`logout` is called.
@@ -1029,6 +1029,11 @@ class Database(common.BaseObject):
             authentication mechanism specific options. To specify the service
             name for GSSAPI authentication pass
             authMechanismProperties='SERVICE_NAME:<service name>'
+
+        .. versionchanged:: 3.5
+           Deprecated. Authenticating multiple users conflicts with support for
+           logical sessions in MongoDB 3.6. To authenticate as multiple users,
+           create multiple instances of MongoClient.
 
         .. versionadded:: 2.8
            Use SCRAM-SHA-1 with MongoDB 3.0 and later.
@@ -1072,7 +1077,10 @@ class Database(common.BaseObject):
         return True
 
     def logout(self):
-        """Deauthorize use of this database for this client instance."""
+        """**DEPRECATED**: Deauthorize use of this database."""
+        warnings.warn("Database.logout() is deprecated",
+                      DeprecationWarning, stacklevel=2)
+
         # Sockets will be deauthenticated as they are used.
         self.client._purge_credentials(self.name)
 
@@ -1101,18 +1109,7 @@ class Database(common.BaseObject):
         return self[dbref.collection].find_one({"_id": dbref.id}, **kwargs)
 
     def eval(self, code, *args):
-        """Evaluate a JavaScript expression in MongoDB.
-
-        Useful if you need to touch a lot of data lightly; in such a
-        scenario the network transfer of the data could be a
-        bottleneck. The `code` argument must be a JavaScript
-        function. Additional positional arguments will be passed to
-        that function when it is run on the server.
-
-        Raises :class:`TypeError` if `code` is not an instance of
-        :class:`basestring` (:class:`str` in python 3) or `Code`.
-        Raises :class:`~pymongo.errors.OperationFailure` if the eval
-        fails. Returns the result of the evaluation.
+        """**DEPRECATED**: Evaluate a JavaScript expression in MongoDB.
 
         :Parameters:
           - `code`: string representation of JavaScript code to be
@@ -1123,6 +1120,9 @@ class Database(common.BaseObject):
         .. warning:: the eval command is deprecated in MongoDB 3.0 and
           will be removed in a future server version.
         """
+        warnings.warn("Database.eval() is deprecated",
+                      DeprecationWarning, stacklevel=2)
+
         if not isinstance(code, Code):
             code = Code(code)
 
@@ -1139,30 +1139,17 @@ class Database(common.BaseObject):
 
 
 class SystemJS(object):
-    """Helper class for dealing with stored JavaScript.
+    """**DEPRECATED**: Helper class for dealing with stored JavaScript.
     """
 
     def __init__(self, database):
-        """Get a system js helper for the database `database`.
+        """**DEPRECATED**: Get a system js helper for the database `database`.
 
-        An instance of :class:`SystemJS` can be created with an instance
-        of :class:`Database` through :attr:`Database.system_js`,
-        manual instantiation of this class should not be necessary.
-
-        :class:`SystemJS` instances allow for easy manipulation and
-        access to server-side JavaScript:
-
-        .. doctest::
-
-          >>> db.system_js.add1 = "function (x) { return x + 1; }"
-          >>> db.system.js.find({"_id": "add1"}).count()
-          1
-          >>> db.system_js.add1(5)
-          6.0
-          >>> del db.system_js.add1
-          >>> db.system.js.find({"_id": "add1"}).count()
-          0
+        SystemJS will be removed in PyMongo 4.0.
         """
+        warnings.warn("SystemJS is deprecated",
+                      DeprecationWarning, stacklevel=2)
+
         if not database.write_concern.acknowledged:
             database = database.client.get_database(
                 database.name, write_concern=WriteConcern())